Экран Stories переписан на MVVM
